chore(test): move and refactor tests
